requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.5.2
astunparse==1.6.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click @ file:///croot/click_1698129812380/work
colorama @ file:///croot/colorama_1672386526460/work
contourpy==1.1.1
cycler==0.12.1
deepblink==0.1.4
docutils==0.20.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.8
ffmpy==0.5.0
filelock @ file:///croot/filelock_1700591183607/work
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2025.2.0
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grad-cam==1.4.8
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.68.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2 @ file:///croot/jinja2_1706733616596/work
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.5
mdurl==0.1.2
mpmath @ file:///croot/mpmath_1690848262763/work
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==2022.4.8
nvidia-cublas-cu117==11.10.1.25
nvidia-cuda-runtime-cu11==2022.4.25
nvidia-cuda-runtime-cu117==11.7.60
nvidia-cudnn-cu11==2022.5.19
nvidia-cudnn-cu116==8.4.0.27
nvidia-ml-py==12.570.86
nvidia-pyindex==1.0.9
nvidia-tensorrt==8.4.3.1
nvitop==1.4.2
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
orjson==3.10.15
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.0.3
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.6
protobuf==4.25.5
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycuda==2024.1.2
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytools==2024.1.14
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
ruff==0.9.6
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.18.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.44.0
statistics==1.0.3.5
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.7.10
tomlkit==0.12.0
torch==1.13.1+cu116
torch2trt @ git+https://github.com/NVIDIA-AI-IOT/torch2trt.git@4e820ae31b4e35d59685935223b05b2e11d47b03
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.67.0
ttach==0.0.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.0.51
urllib3==2.2.3
uvicorn==0.33.0
websocket-client @ file:///croot/websocket-client_1715878298792/work
websockets==12.0
Werkzeug==3.0.6
wrapt==1.17.0
XlsxWriter==3.2.2
zipp==3.20.2
